back/requirements.txt
Flask
kafka-python
python-dotenv
requests
cassandra-driver